+ Merge GLFW2 and GLFW3 input event testing together
